% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{dose_response}
\alias{dose_response}
\title{Dose response for Physical Activity}
\usage{
dose_response(
  cause,
  outcome_type,
  dose,
  quantile = 0.5,
  censor_method = "default",
  confidence_intervals = F
)
}
\arguments{
\item{cause}{All-cause-mortality or disease name}

\item{outcome_type}{Fatal, non-fatal or fatal-and-non-fatal}

\item{dose}{Dose (all-cause or disease)}

\item{quantile}{Numeric value between 0 and 1 - default is 0.5}

\item{censor_method}{Censoring method to use. Options are:
  \describe{
    \item{`none`}{No censoring.}
    \item{`default`}{The default censoring method identified by the meta-analysis.}
    \item{`WHO-DRL`}{A fixed WHO-recommended double level of physical activity
      (17.5 MMET-hours per week).}
    \item{`WHO-QRL`}{A fixed WHO-recommended quadruple level of physical activity
      (35 MMET-hours per week).}
}}

\item{confidence_intervals}{Boolean variable to determine whether confidence intervals are returned or not - default is FALSE}
}
\value{
response for a specific dose (in a data frame)
}
\description{
Based on our meta-analysis, returns a response value for a specific cause or disease. \cr
1. Disease/Cause: Name of the disease/cause - from a specific list\cr
2. Type: Fatal, non-fatal or fatal-and-non-fatal. \cr
}
\details{
At the end, this function returns a response values with or without confidence interval
}
