% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{\%>\%}
\alias{\%>\%}
\alias{dose_response}
\title{Dose response for Physical Acitivity}
\usage{
dose_response(cause, outcome_type, dose, confidence_interval = F)
}
\arguments{
\item{cause}{All-Cause or disease name}

\item{outcome_type}{Mortality or Incidence}

\item{dose}{Dose (all-cause or disease)}

\item{confidence_interval}{Boolean variable to determine whether confidence intervals are returned or not - default false}
}
\value{
response for a specific dose
}
\description{
Based on our meta-analysis, returns a response value for a specific cause or disease. \cr
1. Disease/Cause: Name of the disease/cause - from a specific list\cr
2. Type: Mortality or Incidence. \cr
}
\details{
At the end, this function returns a response values with or without confidence interval
}
