% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{dose_response}
\alias{dose_response}
\title{Dose response for Physical Activity}
\usage{
dose_response(
  cause,
  outcome_type,
  dose,
  quantile = 0.5,
  censor_method = "75thPercentile",
  confidence_intervals = F
)
}
\arguments{
\item{cause}{All-cause-mortality or disease name}

\item{outcome_type}{Fatal, non-fatal or fatal-and-non-fatal}

\item{dose}{Dose (all-cause or disease)}

\item{quantile}{Numeric value between 0 and 1 - default is 0.5}

\item{censor_method}{Use either no censor method, or the 75th percentile by person years,}

\item{confidence_intervals}{Boolean variable to determine whether confidence intervals are returned or not - default is FALSE}
}
\value{
response for a specific dose (in a data frame)
}
\description{
Based on our meta-analysis, returns a response value for a specific cause or disease. \cr
1. Disease/Cause: Name of the disease/cause - from a specific list\cr
2. Type: Fatal, non-fatal or fatal-and-non-fatal. \cr
}
\details{
At the end, this function returns a response values with or without confidence interval
}
